refactor to cold days
